Mise en place Addition & Soustraction. Ajout classe Command pour mise en place Design Pattern Command par la suite

+++ src/main/java/Controllers/Addition.java
@@ -2,7 +2,7 @@ package Controllers;
 
 public class Addition{
 	
-	public int Addition(int a, int b) {
+	public double Addition(double a, double b) {
 		return a + b;
 	}
